requirements.txt
flask==3.0.3
flask-session==0.8.0
python-dotenv==1.0.1
requests==2.32.3
openai==1.30.1
google-auth==2.29.0
google-api-python-client==2.131.0
sendgrid==6.11.0
ipinfo==5.0.1
markdown==3.6
nbformat==5.10.4
